Optimized website for 1080p
I noticed my website hid content behind the videogames on certain pages
I fixed that bug by rescaling and moving content.
This bug only appeared on certain monitors due to the fact that I scripted the website for my monitor not considering I have a 4k screen and when i went to show my dad the website on a family computer that had a 1080p display it looked completely different. I worked out some of the issues and at least made it possible to click the links regardless of the display size.